Cleanup
This comment is just for cleanup purposes.

The next thing I shall work on is a 2-thread system for the beat system. One thread will be for reading the beats, and the other thread will be for executing the beats. This will ensure that neither thread is bottlenecking the other.